test: Extracted fixture to conftest for more general availability
